move scan functions out of hook
